Added conditional building of spec_helper and added puppet stdlib cloning in fixtures

+++ run_build.sh
@@ -1,14 +1,19 @@
-#!/bin/bash --login
-
+#!/bin/bash
+source /etc/profile.d/rvm.sh
 rvm use 1.8.7-p374@udev --create
 bundle install
+presentstat=`gem list | grep puppetlabs_spec_helper| grep 0.4.1.43 | wc -l`
+if [[ ${presentstat} -eq 0 ]]
+then
 git clone https://github.com/puppetlabs/puppetlabs_spec_helper.git
 cd puppetlabs_spec_helper
 rake package:gem
 gem install pkg/puppetlabs_spec_helper-*.gem --ignore-dependencies --no-ri --no-RDoc
-gem uninstall puppetlabs_spec_helper --version=0.4.1
 rm -rf ../puppetlabs_spec_helper/
-cd ../;rake spec
+cd ..
+fi
+gem uninstall puppetlabs_spec_helper --version=0.4.1
+rake spec
 RESULT=$?
 #rvm --force gemset delete 1.8.7-p374@udev
 exit $RESULT
